separate to apply requirements and handle one or multiple project separately
